velocity-sync-engine-(vse)/requirements.txt
python-dotenv>=1.0.1
google-genai>=1.0.0
httpcore==1.0.5
httpx==0.27.2